Use TVar (Map ..) instead of IORef (Map ..)
The original code even used `MVar` instead of `IORef`, but that was a mistake.

We don't use StmContainers as this gives very similar performance to
StmContainers and avoids using couple of extra libraries.

Also in this commit:
* Use `newIORef (error "impossible")` instead of using MVar to be written
immediately. This should perform better and not have the possibility of getting
stuck forever.